requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
atlastk==0.13.4
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
greenlet==3.2.1
h11==0.16.0
holidays==0.71
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
iso8601==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
oauthlib==3.2.2
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
prophet==1.1.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stanio==0.5.1
starlette==0.46.2
threadpoolctl==3.6.0
tortoise==0.1.1
tortoise-orm==0.25.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
